rowlab

Module measurements

Structs

Measurements

An iterator for WeatherStations that yields random measurements for random weather stations.

WeatherStations

The different locations that we want to aggregate measurements for.

Constants

MAGIC_NUMBER 🔒

A very magical number. Used for seeding the random number generator.

STANDARD_DEVIATION 🔒

The standard deviation for the normally-distributed temperatures.

STATIONS_FILE 🔒

The file containing the different weather stations and their average temperatures.

STATIONS_PATTERN 🔒

The regex pattern to read in the possible weather stations and their average temperatures.